Refactor the skill hooks and rolls into base skill data model
