+++ module/documents/items/misc/misc-ability-data-model.mjs
@@ -1,86 +1,8 @@
 import { ProgressDataModel } from '../common/progress-data-model.mjs';
 import { MiscAbilityMigrations } from './misc-ability-migrations.mjs';
-import { CheckHooks } from '../../../checks/check-hooks.mjs';
 import { deprecationNotice } from '../../../helpers/deprecation-helper.mjs';
-import { Checks } from '../../../checks/checks.mjs';
-import { CheckConfiguration } from '../../../checks/check-configuration.mjs';
-import { ChooseWeaponDialog } from '../skill/choose-weapon-dialog.mjs';
-import { CHECK_DETAILS } from '../../../checks/default-section-order.mjs';
-import { CommonSections } from '../../../checks/common-sections.mjs';
 import { ItemPartialTemplates } from '../item-partial-templates.mjs';
-import { TraitUtils } from '../../../pipelines/traits.mjs';
 import { BaseSkillDataModel } from '../skill/base-skill-data-model.mjs';
-import { CommonEvents } from '../../../checks/common-events.mjs';
-
-const skillForAttributeCheck = 'skillForAttributeCheck';
-
-/**
- * @type RenderCheckHook
- */
-let onRenderAccuracyCheck = async (data, check, actor, item, flags) => {
-	if (check.type === 'accuracy' && item?.system instanceof MiscAbilityDataModel) {
-		const inspector = CheckConfiguration.inspect(check);
-		const weapon = await fromUuid(inspector.getWeaponReference());
-		if (check.critical) {
-			CommonSections.opportunity(data.sections, item.system.opportunity, CHECK_DETAILS);
-		}
-		CommonSections.description(data.sections, item.system.description, item.system.summary.value, CHECK_DETAILS);
-		if (item.system.hasClock.value) {
-			CommonSections.clock(data.sections, item.system.progress, CHECK_DETAILS);
-		}
-
-		// Tag info
-		let tags = [];
-		tags.push(...TraitUtils.toTags(item.system.traits));
-		if (weapon) {
-			if (weapon.system.getTags instanceof Function) {
-				tags.push(...weapon.system.getTags(item.system.useWeapon.traits));
-			}
-		}
-		CommonSections.tags(data.sections, tags, CHECK_DETAILS);
-	}
-};
-Hooks.on(CheckHooks.renderCheck, onRenderAccuracyCheck);
-
-/**
- * @type RenderCheckHook
- */
-let onRenderAttributeCheck = async (data, check, actor, item, flags) => {
-	if (check.type === 'attribute' && item?.system instanceof MiscAbilityDataModel && check.additionalData[skillForAttributeCheck]) {
-		const inspector = CheckConfiguration.inspect(check);
-		const ability = await fromUuid(inspector.getWeaponReference());
-		CommonSections.itemFlavor(data.sections, ability);
-		if (check.critical) {
-			CommonSections.opportunity(data.sections, ability.system.opportunity, CHECK_DETAILS);
-		}
-		CommonSections.description(data.sections, ability.system.description, ability.system.summary.value, CHECK_DETAILS, true);
-		if (ability.system.hasClock.value) {
-			CommonSections.clock(data.sections, item.system.progress, CHECK_DETAILS);
-		}
-	}
-};
-
-Hooks.on(CheckHooks.renderCheck, onRenderAttributeCheck);
-
-/**
- * @type RenderCheckHook
- */
-const onRenderDisplay = (data, check, actor, item, flags) => {
-	if (check.type === 'display' && item?.system instanceof MiscAbilityDataModel) {
-		const skill = item.system;
-		CommonSections.tags(data.sections, skill.getCommonTags(), CHECK_DETAILS);
-		if (item.system.hasResource.value) {
-			CommonSections.resource(data.sections, item.system.rp, CHECK_DETAILS);
-		}
-		CommonSections.description(data.sections, item.system.description, item.system.summary.value, CHECK_DETAILS);
-		const inspector = CheckConfiguration.inspect(check);
-		const targets = inspector.getTargetsOrDefault();
-		CommonSections.actions(data, actor, item, targets, flags, inspector);
-		CommonEvents.skill(actor, item);
-	}
-};
-
-Hooks.on(CheckHooks.renderCheck, onRenderDisplay);
 
 /**
  * @property {string} subtype.value
@@ -140,75 +62,8 @@ export class MiscAbilityDataModel extends BaseSkillDataModel {
 		return source;
 	}
 
-	/**
-	 * @param {KeyboardModifiers} modifiers
-	 * @return {Promise<void>}
-	 */
-	async roll(modifiers) {
-		if (this.hasRoll.value) {
-			if (this.useWeapon.accuracy) {
-				return Checks.accuracyCheck(this.parent.actor, this.parent, this.#initializeAccuracyCheck(modifiers));
-			} else {
-				return Checks.attributeCheck(
-					this.parent.actor,
-					{
-						primary: this.attributes.primary,
-						secondary: this.attributes.secondary,
-					},
-					this.parent,
-					this.#initializeAttributeCheck(modifiers),
-				);
-			}
-		}
-		return Checks.display(this.parent.actor, this.parent, this.#initializeSkillDisplay(modifiers));
-	}
-
-	/**
-	 * @param {KeyboardModifiers} modifiers
-	 * @return {CheckCallback}
-	 * @remarks Expects a weapon
-	 */
-	#initializeSkillDisplay(modifiers) {
-		return async (check, actor, item) => {
-			const config = CheckConfiguration.configure(check);
-			await this.configureDisplayCheck(config, actor, item);
-		};
-	}
-
-	/**
-	 * @return {CheckCallback}
-	 */
-	#initializeAttributeCheck(modifiers) {
-		return async (check, actor, item) => {
-			await this.configureAttributeCheck(modifiers, check, actor, item);
-		};
-	}
-
-	/**
-	 * @param modifiers
-	 * @return {CheckCallback}
-	 */
-	#initializeAccuracyCheck(modifiers) {
-		return async (check, actor, item) => {
-			const weapon = await ChooseWeaponDialog.prompt(actor, true);
-			if (weapon === false) {
-				let message = game.i18n.localize('FU.AbilityNoWeaponEquipped');
-				ui.notifications.error(message);
-				throw new Error(message);
-			}
-			if (weapon == null) {
-				throw new Error('no selection');
-			}
-			const { check: weaponCheck, error } = await Checks.prepareCheckDryRun('accuracy', actor, weapon);
-			if (error) {
-				throw error;
-			}
-
-			check.primary = weaponCheck.primary;
-			check.secondary = weaponCheck.secondary;
-
-			return this.configureAccuracyCheck(modifiers, check, actor, item, weapon);
-		};
+	getTags() {
+		return super.getTags();
 	}
 
 	get attributePartials() {
